@@ -5,40 +5,39 @@ using System.Net;
 using System.Threading.Tasks;
 using System;
 
-namespace Basket.API.Controllers
+namespace Basket.API.Controllers;
+
+[ApiController]
+[Route("api/v1/[controller]")]
+public class BasketController : ControllerBase
 {
-    [ApiController]
-    [Route("api/v1/[controller]")]
-    public class BasketController : ControllerBase
+    private readonly IBasketRepository _repository;
+
+    public BasketController(IBasketRepository repository)
     {
-        private readonly IBasketRepository _repository;
+        _repository = repository ?? throw new ArgumentNullException(nameof(repository));
+    }
 
-        public BasketController(IBasketRepository repository)
-        {
-            _repository = repository ?? throw new ArgumentNullException(nameof(repository));
-        }
+    [HttpGet("{userName}", Name = "GetBasket")]
+    [ProducesResponseType(typeof(ShoppingCart), (int)HttpStatusCode.OK)]
+    public async Task<ActionResult<ShoppingCart>> GetBasket(string userName)
+    {
+        var basket = await _repository.GetBasket(userName);
+        return Ok(basket ?? new ShoppingCart(userName));
+    }
 
-        [HttpGet("{userName}", Name = "GetBasket")]
-        [ProducesResponseType(typeof(ShoppingCart), (int)HttpStatusCode.OK)]
-        public async Task<ActionResult<ShoppingCart>> GetBasket(string userName)
-        {
-            var basket = await _repository.GetBasket(userName);
-            return Ok(basket ?? new ShoppingCart(userName));
-        }
+    [HttpPost]
+    [ProducesResponseType(typeof(ShoppingCart), (int)HttpStatusCode.OK)]
+    public async Task<ActionResult<ShoppingCart>> UpdateBasket([FromBody] ShoppingCart basket)
+    {
+        return Ok(await _repository.UpdateBasket(basket));
+    }
 
-        [HttpPost]
-        [ProducesResponseType(typeof(ShoppingCart), (int)HttpStatusCode.OK)]
-        public async Task<ActionResult<ShoppingCart>> UpdateBasket([FromBody] ShoppingCart basket)
-        {
-            return Ok(await _repository.UpdateBasket(basket));
-        }
-
-        [HttpDelete("{userName}", Name = "DeleteBasket")]
-        [ProducesResponseType(typeof(void), (int)HttpStatusCode.OK)]
-        public async Task<IActionResult> DeleteBasket(string userName)
-        {
-            await _repository.DeleteBasket(userName);
-            return Ok();
-        }
+    [HttpDelete("{userName}", Name = "DeleteBasket")]
+    [ProducesResponseType(typeof(void), (int)HttpStatusCode.OK)]
+    public async Task<IActionResult> DeleteBasket(string userName)
+    {
+        await _repository.DeleteBasket(userName);
+        return Ok();
     }
 }
