@@ -1,32 +1,31 @@
 ﻿using System.Collections.Generic;
 
-namespace Basket.API.Entities
+namespace Basket.API.Entities;
+
+public class ShoppingCart
 {
-    public class ShoppingCart
+    public string UserName { get; set; }
+    public List<ShoppingCartItem> Items { get; set; } = new List<ShoppingCartItem>();
+
+    public ShoppingCart()
     {
-        public string UserName { get; set; }
-        public List<ShoppingCartItem> Items { get; set; } = new List<ShoppingCartItem>();
+    }
 
-        public ShoppingCart()
-        {
-        }
+    public ShoppingCart(string userName)
+    {
+        UserName = userName;
+    }
 
-        public ShoppingCart(string userName)
+    public decimal TotalPrice
+    {
+        get
         {
-            UserName = userName;
-        }
-
-        public decimal TotalPrice
-        {
-            get
+            decimal totalprice = 0;
+            foreach (var item in Items)
             {
-                decimal totalprice = 0;
-                foreach (var item in Items)
-                {
-                    totalprice += item.Price * item.Quantity;
-                }
-                return totalprice;
+                totalprice += item.Price * item.Quantity;
             }
+            return totalprice;
         }
     }
 }
