@@ -4,37 +4,36 @@ using System.Threading.Tasks;
 using System;
 using Newtonsoft.Json;
 
-namespace Basket.API.Repositories
+namespace Basket.API.Repositories;
+
+public class BasketRepository : IBasketRepository
 {
-    public class BasketRepository : IBasketRepository
+    private readonly IDistributedCache _redisCache;
+
+    public BasketRepository(IDistributedCache redisCache)
     {
-        private readonly IDistributedCache _redisCache;
+        _redisCache = redisCache ?? throw new ArgumentNullException(nameof(redisCache));
+    }
 
-        public BasketRepository(IDistributedCache redisCache)
-        {
-            _redisCache = redisCache ?? throw new ArgumentNullException(nameof(redisCache));
-        }
+    public async Task<ShoppingCart> GetBasket(string userName)
+    {
+        var basket = await _redisCache.GetStringAsync(userName);
 
-        public async Task<ShoppingCart> GetBasket(string userName)
-        {
-            var basket = await _redisCache.GetStringAsync(userName);
+        if (String.IsNullOrEmpty(basket))
+            return null;
 
-            if (String.IsNullOrEmpty(basket))
-                return null;
+        return JsonConvert.DeserializeObject<ShoppingCart>(basket);
+    }
 
-            return JsonConvert.DeserializeObject<ShoppingCart>(basket);
-        }
+    public async Task<ShoppingCart> UpdateBasket(ShoppingCart basket)
+    {
+        await _redisCache.SetStringAsync(basket.UserName, JsonConvert.SerializeObject(basket));
 
-        public async Task<ShoppingCart> UpdateBasket(ShoppingCart basket)
-        {
-            await _redisCache.SetStringAsync(basket.UserName, JsonConvert.SerializeObject(basket));
+        return await GetBasket(basket.UserName);
+    }
 
-            return await GetBasket(basket.UserName);
-        }
-
-        public async Task DeleteBasket(string userName)
-        {
-            await _redisCache.RemoveAsync(userName);
-        }
+    public async Task DeleteBasket(string userName)
+    {
+        await _redisCache.RemoveAsync(userName);
     }
 }
