@@ -1,12 +1,11 @@
 ﻿using Basket.API.Entities;
 using System.Threading.Tasks;
 
-namespace Basket.API.Repositories
+namespace Basket.API.Repositories;
+
+public interface IBasketRepository
 {
-    public interface IBasketRepository
-    {
-        Task<ShoppingCart> GetBasket(string userName);
-        Task<ShoppingCart> UpdateBasket(ShoppingCart basket);
-        Task DeleteBasket(string userName);
-    }
+    Task<ShoppingCart> GetBasket(string userName);
+    Task<ShoppingCart> UpdateBasket(ShoppingCart basket);
+    Task DeleteBasket(string userName);
 }
