@@ -7,73 +7,72 @@ using System.Collections.Generic;
 using System.Net;
 using System.Threading.Tasks;
 
-namespace Catalog.API.Controllers
+namespace Catalog.API.Controllers;
+
+[ApiController]
+[Route("api/v1/[controller]")]
+public class CatalogController : ControllerBase
 {
-    [ApiController]
-    [Route("api/v1/[controller]")]
-    public class CatalogController : ControllerBase
+    private readonly IProductRepository _repository;
+    private readonly ILogger<CatalogController> _logger;
+
+    public CatalogController(IProductRepository repository, ILogger<CatalogController> logger)
     {
-        private readonly IProductRepository _repository;
-        private readonly ILogger<CatalogController> _logger;
+        _repository = repository ?? throw new ArgumentNullException(nameof(repository));
+        _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+    }
 
-        public CatalogController(IProductRepository repository, ILogger<CatalogController> logger)
+    [HttpGet]
+    [ProducesResponseType(typeof(IEnumerable<Product>), (int)HttpStatusCode.OK)]
+    public async Task<ActionResult<IEnumerable<Product>>> GetProducts()
+    {
+        var products = await _repository.GetProducts();
+        return Ok(products);
+    }
+
+    [HttpGet("{id:length(24)}", Name = "GetProduct")]
+    [ProducesResponseType((int)HttpStatusCode.NotFound)]
+    [ProducesResponseType(typeof(Product), (int)HttpStatusCode.OK)]
+    public async Task<ActionResult<Product>> GetProductById(string id)
+    {
+        var product = await _repository.GetProduct(id);
+        if (product == null)
         {
-            _repository = repository ?? throw new ArgumentNullException(nameof(repository));
-            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+            _logger.LogError($"Product with id: {id}, not found.");
+            return NotFound();
         }
+        return Ok(product);
+    }
 
-        [HttpGet]
-        [ProducesResponseType(typeof(IEnumerable<Product>), (int)HttpStatusCode.OK)]
-        public async Task<ActionResult<IEnumerable<Product>>> GetProducts()
-        {
-            var products = await _repository.GetProducts();
-            return Ok(products);
-        }
+    [Route("[action]/{category}", Name = "GetProductByCategory")]
+    [HttpGet]
+    [ProducesResponseType(typeof(IEnumerable<Product>), (int)HttpStatusCode.OK)]
+    public async Task<ActionResult<IEnumerable<Product>>> GetProductByCategory(string category)
+    {
+        var products = await _repository.GetProductByCategory(category);
+        return Ok(products);
+    }
 
-        [HttpGet("{id:length(24)}", Name = "GetProduct")]
-        [ProducesResponseType((int)HttpStatusCode.NotFound)]
-        [ProducesResponseType(typeof(Product), (int)HttpStatusCode.OK)]
-        public async Task<ActionResult<Product>> GetProductById(string id)
-        {
-            var product = await _repository.GetProduct(id);
-            if (product == null)
-            {
-                _logger.LogError($"Product with id: {id}, not found.");
-                return NotFound();
-            }
-            return Ok(product);
-        }
+    [HttpPost]
+    [ProducesResponseType(typeof(Product), (int)HttpStatusCode.OK)]
+    public async Task<ActionResult<Product>> CreateProduct([FromBody] Product product)
+    {
+        await _repository.CreateProduct(product);
 
-        [Route("[action]/{category}", Name = "GetProductByCategory")]
-        [HttpGet]
-        [ProducesResponseType(typeof(IEnumerable<Product>), (int)HttpStatusCode.OK)]
-        public async Task<ActionResult<IEnumerable<Product>>> GetProductByCategory(string category)
-        {
-            var products = await _repository.GetProductByCategory(category);
-            return Ok(products);
-        }
+        return CreatedAtRoute("GetProduct", new { id = product.Id }, product);
+    }
 
-        [HttpPost]
-        [ProducesResponseType(typeof(Product), (int)HttpStatusCode.OK)]
-        public async Task<ActionResult<Product>> CreateProduct([FromBody] Product product)
-        {
-            await _repository.CreateProduct(product);
+    [HttpPut]
+    [ProducesResponseType(typeof(Product), (int)HttpStatusCode.OK)]
+    public async Task<IActionResult> UpdateProduct([FromBody] Product product)
+    {
+        return Ok(await _repository.UpdateProduct(product));
+    }
 
-            return CreatedAtRoute("GetProduct", new { id = product.Id }, product);
-        }
-
-        [HttpPut]
-        [ProducesResponseType(typeof(Product), (int)HttpStatusCode.OK)]
-        public async Task<IActionResult> UpdateProduct([FromBody] Product product)
-        {
-            return Ok(await _repository.UpdateProduct(product));
-        }
-
-        [HttpDelete("{id:length(24)}", Name = "DeleteProduct")]
-        [ProducesResponseType(typeof(Product), (int)HttpStatusCode.OK)]
-        public async Task<IActionResult> DeleteProductById(string id)
-        {
-            return Ok(await _repository.DeleteProduct(id));
-        }
+    [HttpDelete("{id:length(24)}", Name = "DeleteProduct")]
+    [ProducesResponseType(typeof(Product), (int)HttpStatusCode.OK)]
+    public async Task<IActionResult> DeleteProductById(string id)
+    {
+        return Ok(await _repository.DeleteProduct(id));
     }
 }
