@@ -2,19 +2,18 @@
 using Microsoft.Extensions.Configuration;
 using MongoDB.Driver;
 
-namespace Catalog.API.Data
+namespace Catalog.API.Data;
+
+public class CatalogContext : ICatalogContext
 {
-    public class CatalogContext : ICatalogContext
+    public CatalogContext(IConfiguration configuration)
     {
-        public CatalogContext(IConfiguration configuration)
-        {
-            var client = new MongoClient(configuration.GetValue<string>("DatabaseSettings:ConnectionString"));
-            var database = client.GetDatabase(configuration.GetValue<string>("DatabaseSettings:DatabaseName"));
+        var client = new MongoClient(configuration.GetValue<string>("DatabaseSettings:ConnectionString"));
+        var database = client.GetDatabase(configuration.GetValue<string>("DatabaseSettings:DatabaseName"));
 
-            Products = database.GetCollection<Product>(configuration.GetValue<string>("DatabaseSettings:CollectionName"));
-            CatalogContextSeed.SeedData(Products);
-        }
-
-        public IMongoCollection<Product> Products { get; }
+        Products = database.GetCollection<Product>(configuration.GetValue<string>("DatabaseSettings:CollectionName"));
+        CatalogContextSeed.SeedData(Products);
     }
+
+    public IMongoCollection<Product> Products { get; }
 }
