@@ -1,10 +1,9 @@
 ﻿using Catalog.API.Entities;
 using MongoDB.Driver;
 
-namespace Catalog.API.Data
+namespace Catalog.API.Data;
+
+public interface ICatalogContext
 {
-    public interface ICatalogContext
-    {
-        IMongoCollection<Product> Products { get; }
-    }
-} 
+    IMongoCollection<Product> Products { get; }
+}
