@@ -2,17 +2,16 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
-namespace Catalog.API.Repositories
-{
-    public interface IProductRepository
-    {
-        Task<IEnumerable<Product>> GetProducts();
-        Task<Product> GetProduct(string id);
-        Task<IEnumerable<Product>> GetProductByName(string name);
-        Task<IEnumerable<Product>> GetProductByCategory(string categoryName);
+namespace Catalog.API.Repositories;
 
-        Task CreateProduct(Product product);
-        Task<bool> UpdateProduct(Product product);
-        Task<bool> DeleteProduct(string id);
-    }
+public interface IProductRepository
+{
+    Task<IEnumerable<Product>> GetProducts();
+    Task<Product> GetProduct(string id);
+    Task<IEnumerable<Product>> GetProductByName(string name);
+    Task<IEnumerable<Product>> GetProductByCategory(string categoryName);
+
+    Task CreateProduct(Product product);
+    Task<bool> UpdateProduct(Product product);
+    Task<bool> DeleteProduct(string id);
 }
